requirements.txt
playwright==1.42.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
pandas==2.2.1
